requirements.txt
nuscenes-devkit==1.1.11
numpy<2
pyquaternion
opencv-python==4.8.1.78
matplotlib==3.5.3
Shapely==1.8.5.post1
pycocotools
tqdm
